cleaned up code

+++ conv_neural_network.py
@@ -43,10 +43,10 @@ mnist = tf.keras.datasets.mnist
 # print(train_labels.shape)	# (60000,)
 
 train_images = train_images.reshape((60000, 28, 28, 1))
-train_images = train_images.astype('float32') / 255
+train_images = train_images / 255.0
 
 test_images = test_images.reshape((10000, 28, 28, 1))
-test_images = test_images.astype('float32') / 255
+test_images = test_images / 255.0
 
 train_labels = to_categorical(train_labels)
 test_labels = to_categorical(test_labels)
